feat: remove sample project and configure tests
- Add `Microsoft.AspNetCore.Mvc.NewtonsoftJson`.
- `NewtonsoftJson` is required by the framework.
- Use a new database for each test method to avoid race condition.
- Use SqlServer instead of InMemory database
