+++ AspNetCore.RestFramework.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31424.327
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.RestFramework.Sample", "samples\AspNetCore.RestFramework.Sample\AspNetCore.RestFramework.Sample.csproj", "{1BD03AAE-F48D-4153-8DAD-964E4CD81CD4}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.RestFramework.Core.Test", "tests\AspNetCore.RestFramework.Core.Test\AspNetCore.RestFramework.Core.Test.csproj", "{5C337F78-3254-477E-9937-42FD595F9733}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspNetCore.RestFramework.Core", "src\AspNetCore.RestFramework.Core\AspNetCore.RestFramework.Core.csproj", "{CA09285F-319F-459A-844D-9AF468C0ABF5}"
