--- samples/AspNetCore.RestFramework.Sample/Controllers/CustomersController.cs
@@ -1,41 +1,0 @@
-﻿using AspNetCore.RestFramework.Core.Base;
-using AspNetCore.RestFramework.Core.Filters;
-using Microsoft.AspNetCore.Mvc;
-using System;
-using AspNetRestFramework.Sample.Context;
-using AspNetRestFramework.Sample.CustomSerializers;
-using AspNetRestFramework.Sample.DTO;
-using AspNetRestFramework.Sample.Filters;
-using AspNetRestFramework.Sample.Models;
-using Microsoft.Extensions.Logging;
-
-namespace AspNetRestFramework.Sample.Controllers
-{
-    [Route("api/[controller]")]
-    [ApiController]
-    public class CustomersController : BaseController<CustomerDto, Customer, Guid, ApplicationDbContext>
-    {
-        public CustomersController(
-            CustomerSerializer serializer,
-            ApplicationDbContext dbContext,
-            ILogger<Customer> logger)
-            : base(
-                  serializer,
-                  dbContext,
-                  logger)
-        {
-            AllowedFields = new[] {
-                nameof(Customer.Id),
-                nameof(Customer.Name),
-                nameof(Customer.CNPJ),
-                nameof(Customer.Age),
-            };
-
-            Filters.Add(new QueryStringFilter<Customer>(AllowedFields));
-            Filters.Add(new QueryStringSearchFilter<Customer>(AllowedFields));
-            Filters.Add(new QueryStringIdRangeFilter<Customer, Guid>());
-            Filters.Add(new DocumentFilter());
-            Filters.Add(new CustomerDocumentIncludeFilter());
-        }
-    }
-}
