--- samples/AspNetCore.RestFramework.Sample/DTO/IntAsIdEntityDto.cs
@@ -1,9 +1,0 @@
-﻿using AspNetCore.RestFramework.Core.Base;
-
-namespace AspNetRestFramework.Sample.DTO
-{
-    public class IntAsIdEntityDto : BaseDto<int>
-    {
-        public string Name { get; set; }
-    }
-}
