--- samples/AspNetCore.RestFramework.Sample/DTO/SellerDto.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using AspNetCore.RestFramework.Core.Base;
-
-namespace AspNetRestFramework.Sample.DTO
-{
-    public class SellerDto : BaseDto<Guid>
-    {
-        public string Name { get; set; }
-    }
-}
