--- samples/AspNetCore.RestFramework.Sample/Mappings/CustomerConfig.cs
@@ -1,16 +1,0 @@
-﻿using AspNetRestFramework.Sample.Models;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace AspNetRestFramework.Sample.Mappings
-{
-    public class CustomerConfig : IEntityTypeConfiguration<Customer>
-    {
-        public void Configure(EntityTypeBuilder<Customer> builder)
-        {
-            builder.HasKey(b => b.Id);
-
-            //builder.Navigation(e => e.CustomerDocuments).AutoInclude();
-        }
-    }
-}
