--- samples/AspNetCore.RestFramework.Sample/Mappings/SellerConfig.cs
@@ -1,14 +1,0 @@
-﻿using AspNetRestFramework.Sample.Models;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace AspNetRestFramework.Sample.Mappings
-{
-    public class SellerConfig : IEntityTypeConfiguration<Seller>
-    {
-        public void Configure(EntityTypeBuilder<Seller> builder)
-        {
-            builder.HasKey(b => b.Id);
-        }
-    }
-}
