--- samples/AspNetCore.RestFramework.Sample/Models/IntAsIdEntity.cs
@@ -1,12 +1,0 @@
-﻿using AspNetCore.RestFramework.Core.Base;
-
-namespace AspNetRestFramework.Sample.Models
-{
-    public class IntAsIdEntity : BaseModel<int>
-    {
-        public string Name { get; set; }
-
-        public override string[] GetFields()
-            => new[] { nameof(Id), nameof(Name) };
-    }
-}
