--- samples/AspNetCore.RestFramework.Sample/appsettings.json
@@ -1,14 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information",
-      "Microsoft.EntityFrameworkCore.Database.Command": "Information"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "DefaultConnectionString": "Data Source=localhost,1433;Initial Catalog=RestFramework;User Id=sa;Password=Password1"
-  }
-}
