+++ scripts/start-tests.sh
@@ -2,8 +2,6 @@
 
 set -e
 
-dotnet ef database update --project /app/samples/AspNetCore.RestFramework.Sample/AspNetCore.RestFramework.Sample.csproj
-
 dotnet test tests/AspNetCore.RestFramework.Core.Test \
 --configuration Release \
 --logger trx \
