+++ tests/AspNetCore.RestFramework.Core.Test/Support/Serializers.cs
@@ -1,0 +1,11 @@
+using System;
+using AspNetCore.RestFramework.Core.Serializer;
+
+namespace AspNetCore.RestFramework.Core.Test.Support;
+
+public class CustomerSerializer : Serializer<CustomerDto, Customer, Guid, AppDbContext>
+{
+    public CustomerSerializer(AppDbContext applicationDbContext) : base(applicationDbContext)
+    {
+    }
+}
